feature/add-stop-to-shutdown-handler fix not refactored imports
